[kernel] E: Move KcallArgs to Its Own Module
